pyproject.toml
[tool.poetry]
name = "bloomsage-backend"
version = "0.1.0"
description = ""
authors = ["dolelongan <dolelongan@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,<3.12"
fastapi = "0.103.2"
uvicorn = "0.23.2"
httpx = "0.25.0"
pillow = "10.1.0"
python-multipart = "0.0.6"
torch = "2.1.0"
transformers = "4.35.0"
poethepoet = "0.24.4"
annotated-types = "0.6.0"
anyio = "3.7.1"
build = "1.0.3"
cachecontrol = "0.13.1"
certifi = "2023.7.22"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "7.1.2"
coverage = "7.3.2"
crashtest = "0.4.1"
cryptography = "41.0.7"
distlib = "0.3.7"
dulwich = "0.21.7"
fastjsonschema = "2.19.0"
filelock = "3.13.1"
fsspec = "2023.10.0"
h11 = "0.14.0"
httpcore = "0.18.0"
httptools = "0.6.1"
huggingface-hub = "0.17.3"
idna = "3.4"
importlib-metadata = "7.0.0"
iniconfig = "2.0.0"
installer = "0.7.0"
jaraco-classes = "3.3.0"
jeepney = "0.8.0"
jinja2 = "3.1.2"
keyring = "24.3.0"
markupsafe = "2.1.3"
more-itertools = "10.1.0"
mpmath = "1.3.0"
msgpack = "1.0.7"
mypy = "1.6.1"
mypy-extensions = "1.0.0"
networkx = "3.2.1"
numpy = "1.23.5"
packaging = "23.2"
pastel = "0.2.1"
pathspec = "0.11.2"
pexpect = "4.9.0"
pkginfo = "1.9.6"
platformdirs = "3.11.0"
pluggy = "1.3.0"
poetry = "1.7.1"
poetry-core = "1.8.1"
poetry-plugin-export = "1.6.0"
ptyprocess = "0.7.0"
pycparser = "2.21"
pydantic = "2.4.2"
pydantic-core = "2.10.1"
pyproject-hooks = "1.0.0"
pytest = "7.4.3"
python-dotenv = "1.0.0"
pyyaml = "6.0.1"
rapidfuzz = "3.5.2"
regex = "2023.10.3"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
ruff = "0.1.3"
safetensors = "0.4.0"
secretstorage = "3.3.3"
shellingham = "1.5.4"
sniffio = "1.3.0"
starlette = "0.27.0"
sympy = "1.12"
tokenizers = "0.14.1"
tomli = "2.0.1"
tomlkit = "0.12.3"
tqdm = "4.66.1"
trove-classifiers = "2023.11.29"
typing-extensions = "4.8.0"
urllib3 = "2.0.7"
uvloop = "0.19.0"
virtualenv = "20.25.0"
watchfiles = "0.21.0"
websockets = "12.0"
zipp = "3.17.0"
absl-py = "2.0.0"
astunparse = "1.6.3"
cachetools = "5.3.2"
flatbuffers = "23.5.26"
gast = "0.4.0"
google-auth = "2.25.2"
google-auth-oauthlib = "1.0.0"
google-pasta = "0.2.0"
grpcio = "1.60.0"
h5py = "3.10.0"
jax = "0.4.21"
joblib = "1.3.2"
keras = "2.12.0"
libclang = "16.0.6"
markdown = "3.5.1"
ml-dtypes = "0.3.1"
nvidia-cublas-cu12 = "12.1.3.1"
nvidia-cuda-cupti-cu12 = "12.1.105"
nvidia-cuda-nvrtc-cu12 = "12.1.105"
nvidia-cuda-runtime-cu12 = "12.1.105"
nvidia-cudnn-cu12 = "8.9.2.26"
nvidia-cufft-cu12 = "11.0.2.54"
nvidia-curand-cu12 = "10.3.2.106"
nvidia-cusolver-cu12 = "11.4.5.107"
nvidia-cusparse-cu12 = "12.1.0.106"
nvidia-nccl-cu12 = "2.18.1"
nvidia-nvjitlink-cu12 = "12.3.101"
nvidia-nvtx-cu12 = "12.1.105"
oauthlib = "3.2.2"
opt-einsum = "3.3.0"
pandas = "2.1.4"
protobuf = "4.25.1"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
python-dateutil = "2.8.2"
pytz = "2023.3.post1"
req2toml = "1.2.0"
requests-oauthlib = "1.3.1"
rsa = "4.9"
scikit-learn = "1.2.2"
scipy = "1.11.4"
six = "1.16.0"
tensorboard = "2.12.3"
tensorboard-data-server = "0.7.2"
tensorflow = "2.12.0"
tensorflow-addons = "0.20.0"
tensorflow-estimator = "2.12.0"
tensorflow-io-gcs-filesystem = "0.34.0"
termcolor = "2.4.0"
threadpoolctl = "3.2.0"
triton = "2.1.0"
typeguard = "2.13.3"
tzdata = "2023.3"
werkzeug = "3.0.1"
wrapt = "1.14.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.2"
mypy = "^1.6.1"
pytest = "^7.4.3"
coverage = "^7.3.2"
pillow = "^10.1.0"
python-multipart = "^0.0.6"
torch = "^2.1.0"
transformers = "^4.35.0"

[tool.black]
line-length = 88

[tool.ruff]
select = ["E", "F", "I"]
fixable = ["ALL"]
exclude = [".git", ".mypy_cache", ".ruff_cache"]
line-length = 88

[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = true
plugins = "pydantic.mypy"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.poe.tasks]
compose = "uvicorn 'main:app' --host=0.0.0.0 --port=8000"
lint = "ruff --output-format=github ./main.py"
typecheck = "mypy ./*.py"
test = "run coverage run --source ./ -m pytest"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"